+++ NEWS.md
@@ -24,6 +24,7 @@
 * Fixed problems for `str()` on a corpus with no docvars (#571).
 * `removeURL` in `tokens()` now removes URLs where the first part of the URL is a single letter (#587).
 * `dfm_select` now works correctly for ngram features (#589).
+* Fixed a bug crashing corpus constructors for character vectors with duplicated names (the cause of #580).
 
 
 ## Changes since v0.9.9-17